refactor(bootstrap): マイグレーションロジックを migrations.ts に分離
syncMcCheckReminder, removeLegacyConsolidateReminder, ltm→memory 移行を
apps/discord/src/migrations.ts に抽出し、bootstrap.ts からは import して呼び出すだけにする。
